Visible triggers and tree optimizations.
Added dev materials to be used for marking triggers in general, remember to set triggers to the dev layer so they are not seen in game.

Optimized trees by making tree groups (require less draw calls, less objects to load, less colliders, easier to place in mass), not that the groups dont have colliders and should only be placed outside of player's reach.